@@ -17,8 +17,8 @@
     "issues": "https://github.com/chesio/bc-page-redirect/issues"
   },
   "require": {
-    "php": ">=7.0.0",
-    "composer/installers": "~1.0"
+    "php": "^8.1",
+    "composer/installers": "^1.0 || ^2.0"
   },
   "require-dev": {
     "squizlabs/php_codesniffer": "^3.2"
@@ -33,5 +33,10 @@
     "test": [
       "phpcs"
     ]
+  },
+  "config": {
+    "allow-plugins": {
+      "composer/installers": true
+    }
   }
 }
