@@ -1,7 +1,6 @@
 <?php
-/**
- * @package BC_Page_Redirect
- */
+
+declare(strict_types=1);
 
 namespace BlueChip\PageRedirect\RedirectTypes;
 
@@ -51,7 +50,7 @@ class FirstSubpage extends \BlueChip\PageRedirect\AbstractRedirect
     protected function sanitize(array $data): array
     {
         return [
-            'page_id' => isset($data['page_id']) ? intval($data['page_id']) : 0,
+            'page_id' => intval($data['page_id'] ?? 0),
         ];
     }
 
