@@ -1,7 +1,6 @@
 <?php
-/**
- * @package BC_Page_Redirect
- */
+
+declare(strict_types=1);
 
 namespace BlueChip\PageRedirect;
 
@@ -10,22 +9,34 @@ class Frontend
     /**
      * Initialize front-end integration.
      */
-    public function init()
+    public function init(): void
     {
-        add_action('template_redirect', [$this, 'redirect']);
+        add_action('send_headers', $this->redirect(...), 10, 0);
 
-        add_filter('wp_sitemaps_posts_query_args', [$this, 'filterSitemapQueryArgs'], 10, 1);
+        add_filter('wp_sitemaps_posts_query_args', $this->filterSitemapQueryArgs(...), 10, 1);
     }
 
 
     /**
      * Do a redirect for current request, if applicable, and exit.
-     * @hook https://developer.wordpress.org/reference/hooks/template_redirect/
+     *
+     * @hook https://developer.wordpress.org/reference/hooks/send_headers/
      */
-    public function redirect()
+    private function redirect()
     {
         // Redirect only works with pages.
-        if (is_page() && !empty($location = Core::getRedirectLocation(get_the_ID())) && wp_redirect($location, 301)) {
+        if (!is_page()) {
+            return;
+        }
+
+        $location = Core::getRedirectLocation(get_the_ID());
+
+        // Redirect needs to have a target location set.
+        if ($location === '') {
+            return;
+        }
+
+        if (wp_redirect($location, 301)) {
             exit;
         }
     }
@@ -36,7 +47,7 @@ class Frontend
      *
      * @hook https://developer.wordpress.org/reference/hooks/wp_sitemaps_posts_query_args/
      */
-    public function filterSitemapQueryArgs(array $query_args): array
+    private function filterSitemapQueryArgs(array $query_args): array
     {
         if ($query_args['post_type'] === 'page') {
             // See: https://developer.wordpress.org/reference/classes/WP_Query/#custom-field-post-meta-parameters
